Change new domains when branching. Create filterDomains2
A lot of time was spent in creating new domains on every level,
this process is streamlined a little here by inserting the new
domain in the outer loop and brnaching on the second to las domain.
Have to ensure the extra domain does not break filters and
calculateCost (since we dont have an extra vertex_in_layer yet).
Change has only been made to the restriction method, not relaxation.

filterDomains2 only eliminates singletons.

No significant speedup.