Change calculateCosts, make edges_to_check computation global
Given the current ordering, the edges were remapped every time that
calculateCosts was called. If we assume that the ordering doesn't
change for different nodes this is unneccessary, use a global
edges_to check and just add the new ones when we get to a new level.

Need to confirm this doesnt break subsequent iteration when we use
nodes from the pool as roots.