Change filter to only once. Changes filtering and costs.
Only want to filter a node once, to do this i first add new
domain, then filter. This means that this new domain must be ignored
when computing the costs. And ignored when on last layer and we are
filtering for infeasible (it will always have an empty domain).

Some other small changes.