Change mergeLayer. For domain, break when everything.
mergeLayer used to be for node in layer, for domain. so it happened that
once layer i is everything, we still go to that layer for later nodes.
The change makes it so that the outer loop is the domain i, once it is
full we break and go do domain i+1. ~16% improvement.

Also disabled addbranchNode, since we are only focussing on the root
node currently. No real influence on time.