Change IBS to only branch to children with possibly correct cost.
Usethe current lowerbound as an upperbound for the purpose of domain
pruning before brancing, this leads to fewer nodes that are just created
and immediately cut.

If the loewr bound is increased in this iteration, the same node will
have more children in the next iteration. The count_node statements have
been placed to accurately reflect the eventual total number of nodes in
the tree.

This version of the code was used for the experiments in exp30_1mar16.