Add node limit to the greedy IBS algorithm. BSRelaxation_nodelimit.
Instead of limiting by number of nodes branched on a layer, limit by
total number of nodes in tree. Finish branching the current node when
you cross threshold, then stop.

Limit is imposed on nodes created, not explored (but can be easily
modified).

This version is used for exp30_18mar. (No restriction)